+++ src/utils.rs
@@ -2,11 +2,11 @@ use nalgebra::Vector2;
 use sdl2::rect::Rect;
 
 pub(crate) trait CenterOnVector {
-    fn center_on_vector(&mut self, vector: Vector2<f64>);
+    fn center_on_vector(&mut self, vector: Vector2<i32>);
 }
 
 impl CenterOnVector for Rect {
-    fn center_on_vector(&mut self, vector: Vector2<f64>) {
-        self.center_on((vector.x as i32, vector.y as i32));
+    fn center_on_vector(&mut self, vector: Vector2<i32>) {
+        self.center_on((vector.x, vector.y));
     }
 }
